pyproject.toml
[project]
name = "azure-data-lake-brief-2-04122024"
version = "0.1.0"
description = ""
authors = [
    {name = "dai"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material-extensions (>=1.3.1,<2.0.0)",
    "mkdocs-material (>=9.5.50,<10.0.0)",
    "azure-identity (>=1.19.0,<2.0.0)",
    "azure-keyvault-secrets (>=4.9.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "azure-storage-file-datalake (>=12.18.1,<13.0.0)",
    "azure-storage-blob (>=12.24.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
